% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_general.R
\name{add_publications_authors}
\alias{add_publications_authors}
\title{add_publications_authors}
\usage{
add_publications_authors(n_authors, path = NULL)
}
\arguments{
\item{n_authors}{Number of authors whose publications have to be scraped}

\item{path}{Path ot the current working directory}
}
\value{
A sample of the dataset which has been scrapped
}
\description{
This function automatially update the current dataset "publications_authors" by geting new information on non scrapped authors
}
