% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{as_tibble_c}
\alias{as_tibble_c}
\title{get_publication_by_author}
\usage{
as_tibble_c(x)
}
\arguments{
\item{x}{a list of named elements}
}
\value{
a tibble whose names have been cleaned by janitor::clean_names()
}
\description{
get_publication_by_author
}
