% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{check_attached}
\alias{check_attached}
\title{Check attached packages}
\usage{
check_attached(packages = NULL)
}
\arguments{
\item{packages}{(NULL or Character)}
}
\value{
If \code{packages = NULL}, return a list of currently-attached packages. If
\code{packages} contains a character vector of package names, return a named logical
vector showing if the packages are attached or not.
}
\description{
Check attached packages
}
\examples{
\dontrun{
check_attached()

#> [1] "stats"  "graphics"  "grDevices"  ...

check_attached(c("utils", "stats"))

#> utils stats 
#> TRUE  TRUE 
}

}
