% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{check_installed}
\alias{check_installed}
\title{Check if packages are installed}
\usage{
check_installed(...)
}
\arguments{
\item{...}{(Dots) Package names as bare names, strings, or a character vector.
If left empty, lists all installed packages.}
}
\value{
If \code{dots} is empty, a character vector of all installed packages.
Otherwise, return a named logical vector where \code{TRUE} means the package
is installed.
}
\description{
Check if packages are installed
}
\examples{
\dontrun{
check_installed()

#>   [1] "addinslist"  "antiword" " ape"  "assertthat"  ...

check_installed(c("utils", "stats"))

#> utils stats 
#> TRUE  TRUE 

check_installed("datasets", "base", fakepkg)

#> datasets     base  fakepkg 
#>     TRUE     TRUE    FALSE 
}

}
