% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{collapse_vec}
\alias{collapse_vec}
\title{Collapse a vector}
\usage{
collapse_vec(..., wrap = "'", collapse = ", ", unique = TRUE)
}
\arguments{
\item{...}{(...) Vectors that will be concatenated and coerced to Character.}

\item{wrap}{(Character) Placed at the left and right sides of each vector element.}

\item{collapse}{(Character) Placed between each element of the original vector(s).}

\item{unique}{(Logical) If \code{TRUE}, duplicate entries in \code{...} will be removed.}
}
\value{
A string.
}
\description{
I use this internally for turning a vector of package names into a
}
\examples{
\dontrun{
collapse_vec(month.abb)
#> [1] "'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'"
}

}
