% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dots.R
\name{dots_length}
\alias{dots_length}
\title{How many items are in dots?}
\usage{
dots_length(...)
}
\arguments{
\item{...}{(Dots)}
}
\value{
An integer
}
\description{
How many items are in dots?
}
\examples{
\dontrun{
dots_length(package, names, here)

#> [1] 3
}

}
