% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{eval_quietly}
\alias{eval_quietly}
\title{Suppress warnings/messages/package startup messages}
\usage{
eval_quietly(expr)
}
\arguments{
\item{expr}{An expression.}
}
\value{
Evaluate \code{expr} but suppress warnings/messages/package startup messages.
}
\description{
Suppress warnings/messages/package startup messages
}
