% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{exists_on_CRAN}
\alias{exists_on_CRAN}
\title{Check if a package exists on CRAN}
\usage{
exists_on_CRAN(packages)
}
\arguments{
\item{packages}{(Character) A vector of package names.}
}
\value{
A named logical vector where the names are the URLs that were accessed, and the
result is whether the URL exists (\code{TRUE}) or returns an error (\code{FALSE}).
}
\description{
Most authors check the existence of a CRAN package by using \code{utils::available.packages()}
to fetch a list of all packages that are on CRAN. However, this list is >1 Mb and
'librarian' has empathy for people with bad internet connectivity. Instead, this
function tries to access the CRAN page for each requested package. Pages that return
errors indicate that the package does not exist on CRAN.
}
\examples{
\dontrun{
exists_on_CRAN(c("MASS", "ape", "my_fake_package"))
}

}
