% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{fuzzy_needle}
\alias{fuzzy_needle}
\title{Turn a list of words into a fuzzy regex}
\usage{
fuzzy_needle(vec)
}
\arguments{
\item{vec}{(Character) A string containing space-separated keywords to search for.}
}
\value{
A string where each word has been wrapped as a lookaround term.
}
\description{
A fuzzy regex is one that will match search terms in any order by using PERL 
lookaround. This is very slow, but often worth the cost to get more complete
results.
}
\examples{
\dontrun{
fuzzy_needle("network centrality")
#> [1] "(?=.*network)(?=.*centrality)"
}
}
