% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{is_dots_empty}
\alias{is_dots_empty}
\title{Did the user pass arguments inside dots?}
\usage{
is_dots_empty(...)
}
\arguments{
\item{...}{(Dots)}
}
\value{
\code{TRUE} (dots is empty) or \code{FALSE} (dots is not empty).
}
\description{
Did the user pass arguments inside dots?
}
\examples{
\dontrun{
is_dots_empty(package, names, here)

#> [1] FALSE  
}

}
