% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{is_valid_url}
\alias{is_valid_url}
\title{Assert that a URL is complete and valid}
\usage{
is_valid_url(string)
}
\arguments{
\item{string}{(Character) A URL to check.}
}
\value{
A logical value, \code{TRUE} if the URL is valid, \code{FALSE} if otherwise.
}
\description{
Assert that a URL is complete and valid
}
\details{
The regex I use is "@stephenhay" from
\url{https://mathiasbynens.be/demo/url-regex} because it's the shortest regex that
matches every CRAN mirror at \url{https://cran.r-project.org/mirrors.html}.
}
\examples{
\dontrun{
is_valid_url("http://rstudio.com")
}

}
