% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_status.R
\name{list_dependencies}
\alias{list_dependencies}
\title{List the dependencies of selected packages}
\usage{
list_dependencies(of_pkgs, which = c("Depends", "Imports"))
}
\arguments{
\item{of_pkgs}{(Character) Packages whose dependencies will be found.}

\item{which}{(Character) The types of dependencies to find.}
}
\value{
A character vector of package names. Note that all dependencies of 
   all requested packages will be placed into the one vector.
}
\description{
List the dependencies of selected packages
}
