% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{nse_dots}
\alias{nse_dots}
\title{Non-standard evaluation of dots with base R}
\usage{
nse_dots(..., keep_user = FALSE)
}
\arguments{
\item{...}{(Dots) A dots list.}

\item{keep_user}{(Logical) If \code{FALSE}, remove the Username/ component from GitHub
"Username/Packagename".}
}
\value{
The dots list as a character vector.
}
\description{
Non-standard evaluation of dots with base R
}
\examples{
\donttest{
nse_dots(package, names, here)

#> [1] "package" "names"   "here"   
}

}
