% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{reshelf}
\alias{reshelf}
\title{Detach and then reattach packages}
\usage{
reshelf(...)
}
\arguments{
\item{...}{(Names) Packages as bare names. For packages that come from GitHub, you can
keep the username/package format, or omit the username and provide just the package
name.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Detach and then reattach packages
}
\examples{
# reshelf(desiderata)

}
