% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_author.R
\name{scopus_get_authors}
\alias{scopus_get_authors}
\title{scopus_get_authors}
\usage{
scopus_get_authors(ids = NULL, n_batchs = 99, batch_size = 25,
  api_key = NULL)
}
\arguments{
\item{ids}{vector of ids author ids to scrape}

\item{n_batchs}{number of batch to scrape}

\item{batch_size}{size of scraped batch. Max 25}

\item{api_key}{api_key for Scopus API}
}
\value{
response is a list with two elements. scrapped_ids a vector with the scrapped ids. content is the parsed json response.
}
\description{
scopus_get_authors
}
