% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_publication.R
\name{scopus_get_publication}
\alias{scopus_get_publication}
\title{scopus_get_publication}
\usage{
scopus_get_publication(data)
}
\arguments{
\item{data}{tibble with a column scopus id}
}
\value{
entries, is a list containing the publications of each scrapped author id
}
\description{
scopus_get_publication
}
