% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_author_publication.R
\name{scopus_get_publication_by_author}
\alias{scopus_get_publication_by_author}
\title{scopus_get_publication_by_author}
\usage{
scopus_get_publication_by_author(data)
}
\arguments{
\item{data}{tibble with a column scopus id}
}
\value{
entries, is a list containing the publications of each scrapped author id
}
\description{
scopus_get_publication_by_author
}
