% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_output.R
\name{sentence}
\alias{sentence}
\title{Keep the first sentence of a string.}
\usage{
sentence(string)
}
\arguments{
\item{string}{(Character) A string.}
}
\value{
The string with only the first sentence.
}
\description{
Keep the first sentence of a string.
}
\examples{
\dontrun{
sentence("This is a sentence. And this is another sentence.")

#> [1] "This is a sentence."

sentence("This is just one sentence.")

#> [1] "This is just one sentence."

sentence("Is this a sentence? Or is this one. Maybe this one! What if there are lots of sentences?")

#> [1] "Is this a sentence?
}

}
