% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{shelf}
\alias{shelf}
\title{Attach packages to the search path, installing them from CRAN or GitHub if needed}
\usage{
shelf(..., update_all = FALSE, quiet = FALSE, custom_repo = FALSE)
}
\arguments{
\item{...}{(Names) Packages as bare names. If the package is from GitHub,
include both the username and package name as UserName/package (see examples).}

\item{update_all}{(Logical) If \code{TRUE}, the packages will be re-installed even if they
are already in your library.}

\item{quiet}{(Logical) Suppresses most warnings and messages.}

\item{custom_repo}{(\code{FALSE} or Character) Use \code{FALSE} for the default mirror (in
RStudio you can set a default mirror via \emph{Options > Packages > Default CRAN Mirror}).
Otherwise, provide the URL to a CRAN mirror (e.g. "https://cran.csiro.au/").}
}
\value{
Invisibly returns a named logical vector, where the names are the packages
requested in \code{...} and \code{TRUE} means that the package was successfully installed
and attached.
}
\description{
Attach packages to the search path, installing them from CRAN or GitHub if needed
}
\examples{
# shelf(janitor, DesiQuintans/desiderata, purrr)

# shelf() returns invisibly; bind its output to a variable or access the .Last.value.

# print(.Last.value)

#> janitor desiderata      purrr 
#>    TRUE       TRUE       TRUE 

}
