% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{shhh}
\alias{shhh}
\title{Suppresses console output, including printing}
\usage{
shhh(expr)
}
\arguments{
\item{expr}{(Expression) An expression to evaluate.}
}
\value{
Evaluates \code{expr}.
}
\description{
This is copied from my personal package, \code{desiderata}.
}
