% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{suppress_lib_message}
\alias{suppress_lib_message}
\title{Suppress "lib unspecified" message}
\usage{
suppress_lib_message(expr)
}
\arguments{
\item{expr}{(Expression) A function call.}
}
\value{
Runs the expression, suppressing the "(as 'lib' is unspecified)" message only.
}
\description{
This is used to suppress a specific warning message that is printed by install.packages
and remove.packages, which is caused by the 'lib' arg not being assigned in the 
function call. In particular, devtools::install_github() ultimately
}
\examples{
\dontrun{
suppress_lib_message(remove.packages(fortunes))
}
}
