% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_output.R
\name{tell_user}
\alias{tell_user}
\title{Messages for the user}
\usage{
tell_user(message, ...)
}
\arguments{
\item{message}{(Character) An identifier string for a message.}

\item{...}{(Dots) Data to pass into the message for \code{sprintf()}.}
}
\value{
A string.
}
\description{
Messages for the user
}
\examples{
\dontrun{
message(tell_user("not allowed to make path", "C:/fakefolder"))
}

}
