% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{tibblize_columns}
\alias{tibblize_columns}
\title{tibblize_columns}
\usage{
tibblize_columns(x)
}
\arguments{
\item{x}{a list whoce non list columns have to become lists}
}
\value{
a tibble wherer each column is a list
}
\description{
tibblize_columns
}
