% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{unlist_tibble}
\alias{unlist_tibble}
\title{unlist_tibble}
\usage{
unlist_tibble(x)
}
\arguments{
\item{x}{a list or a tibble}
}
\value{
a list
}
\description{
This function check whether the input is a list on which we could map. If this is not the case, it encapsualit the input in a list.
}
