% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{unshelf}
\alias{unshelf}
\title{Detach (unload) packages from the library}
\usage{
unshelf(...)
}
\arguments{
\item{...}{(Names) Packages as bare names. If a package comes from GitHub, omit the
username and provide just the package name.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Detach (unload) packages from the library
}
\examples{
# unshelf(janitor, desiderata, purrr)

}
