% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{check_installed}
\alias{check_installed}
\title{Check installed packages}
\usage{
check_installed(packages = NULL)
}
\arguments{
\item{packages}{(NULL or Character)}
}
\value{
If \code{packages = NULL}, return a character vector of installed packages. If
\code{packages} contains a character vector of package names, return a named logical
vector showing if the packages are installed or not.
}
\description{
Check installed packages
}
\examples{
\dontrun{
check_installed()

#>   [1] "addinslist"  "antiword" " ape"  "assertthat"  ...

check_installed(c("utils", "stats"))

#> utils stats 
#> TRUE  TRUE 
}

}
