% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{reshelf}
\alias{reshelf}
\title{Detach and then reattach packages to the search path}
\usage{
reshelf(...)
}
\arguments{
\item{...}{(Names) Packages as bare names. For packages that come from GitHub, you can
keep the username/package format, or omit the username and provide just the package
name.}
}
\value{
Invisibly returns a named logical vector, where the names are the packages
requested in \code{...} and \code{TRUE} means that the package was successfully attached.
}
\description{
Detach and then reattach packages to the search path
}
\examples{
\donttest{
reshelf(datasets)

# reshelf() returns invisibly; bind its output to a variable or access the .Last.value.

print(.Last.value)

#> datasets 
#>     TRUE
}

}
