% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{shelf}
\alias{shelf}
\title{Attach packages to the library, installing them from CRAN/GitHub if needed}
\usage{
shelf(..., update_all = FALSE, quiet = FALSE)
}
\arguments{
\item{...}{(Names) Packages as bare names. If the package is from GitHub,
include both the username and package name (see examples).}

\item{update_all}{(Logical) If \code{TRUE}, the packages will be re-installed even if they
already exist in your computer.}

\item{quiet}{(Logical) Suppresses most warnings and messages.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Attach packages to the library, installing them from CRAN/GitHub if needed
}
\examples{
shelf(janitor, DesiQuintans/desiderata, purrr)

}
