% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{shelf}
\alias{shelf}
\title{Attach packages to the library, installing them from CRAN/GitHub if needed}
\usage{
shelf(..., update_all = FALSE, quiet = FALSE,
  repo = "https://cran.rstudio.com/")
}
\arguments{
\item{...}{(Names) Packages as bare names. If the package is from GitHub,
include both the username and package name (see examples).}

\item{update_all}{(Logical) If \code{TRUE}, the packages will be re-installed even if they
already exist in your computer.}

\item{quiet}{(Logical) Suppresses most warnings and messages.}

\item{repo}{(Character) URL to a package repository. By default, the RStudio CRAN
mirror. RStudio can also set a default mirror via \emph{Options > Packages > Default
CRAN Mirror}, and you can use that access that setting
with \code{repo = getOption("repos")}.}
}
\value{
Invisibly returns \code{devtools::session_info()}, which provides version info about
your R and package installations.
}
\description{
Attach packages to the library, installing them from CRAN/GitHub if needed
}
\examples{
shelf(janitor, DesiQuintans/desiderata, purrr)

}
