% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{shelf}
\alias{shelf}
\title{Attach packages to the library, installing them from CRAN/GitHub if needed}
\usage{
shelf(..., update_all = FALSE, quiet = FALSE, custom_repo = FALSE)
}
\arguments{
\item{...}{(Names) Packages as bare names. If the package is from GitHub,
include both the username and package name as UserName/package (see examples).}

\item{update_all}{(Logical) If \code{TRUE}, the packages will be re-installed even if they
already exist in your computer.}

\item{quiet}{(Logical) Suppresses most warnings and messages.}

\item{custom_repo}{(\code{FALSE} or Character) Use \code{FALSE} for the default mirror (in
RStudio you can set a default mirror via \emph{Options > Packages > Default CRAN Mirror}).
Otherwise, provide the URL to a CRAN mirror (e.g. "https://cran.csiro.au/").}
}
\value{
Invisibly returns \code{devtools::session_info()}, which provides version info about
your R and package installations.
}
\description{
Attach packages to the library, installing them from CRAN/GitHub if needed
}
\examples{
# shelf(janitor, DesiQuintans/desiderata, purrr)

}
