% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{unshelf}
\alias{unshelf}
\title{Detach (unload) packages from the library}
\usage{
unshelf(...)
}
\arguments{
\item{...}{(Names) Packages as bare names. For packages that come from GitHub, you can
keep the username/package format, or omit the username and provide just the package
name.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Detach (unload) packages from the library
}
\examples{
# These are the same:

# unshelf(janitor, desiderata, purrr)
# unshelf(janitor, DesiQuintans/desiderata, purrr)

# You can quickly unload-reload packages by just changing 'shelf' to 'unshelf'.

#   shelf(janitor, DesiQuintans/desiderata, purrr)
# unshelf(janitor, DesiQuintans/desiderata, purrr)

}
